+++ lib/l10n/app_localizations.dart
@@ -1057,6 +1057,30 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'Automatically scroll the list when MIDI is received'**
   String get synthAutoScrollSubtitle;
+
+  /// No description provided for @vocoderWarningTitle.
+  ///
+  /// In en, this message translates to:
+  /// **'Feedback Warning'**
+  String get vocoderWarningTitle;
+
+  /// No description provided for @vocoderWarningBody.
+  ///
+  /// In en, this message translates to:
+  /// **'Using the internal microphone and speakers simultaneously can cause a loud feedback loop (Larsen effect). Please use external headphones, a separate microphone, or an external speaker for a safe experience.'**
+  String get vocoderWarningBody;
+
+  /// No description provided for @vocoderWarningValidate.
+  ///
+  /// In en, this message translates to:
+  /// **'Enable Vocoder'**
+  String get vocoderWarningValidate;
+
+  /// No description provided for @vocoderWarningCancel.
+  ///
+  /// In en, this message translates to:
+  /// **'Cancel'**
+  String get vocoderWarningCancel;
 }
 
 class _AppLocalizationsDelegate